requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
coveralls==1.4.0
cryptography==2.3.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-extensions==2.1.2
djangorestframework==3.8.2
docopt==0.6.2
gunicorn==19.9.0
idna==2.7
kombu==4.2.1
oauth2client==4.1.2
oauthlib==2.1.0
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyJWT==1.6.4
pyOpenSSL==18.0.0
python3-openid==3.1.0
pytz==2018.5
honcho==1.0.1
httplib2==0.11.3
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
whitenoise==4.0
